+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## [1.14.4 - 3.2.4]
 
+* fix: correct redstone update
 * ref: update buildpipeline
 
 ## [1.14.4 - 3.2.3]